Sign up and login now work, although the UI still does not reflect the change.

+++ persistence/src/main/java/ar/edu/itba/paw/persistence/APPropertyDao.java
@@ -25,7 +25,7 @@ public class APPropertyDao extends APDao<Property> implements PropertyDao {
             .withCapacity(rs.getInt("capacity"))
             .withCaption(rs.getString("caption"))
             .withDescription(rs.getString("description"))
-            .withNeighbourhoodId(rs.getLong("neighbourhoodId"))
+            .withNeighbourhoodId(rs.getInt("neighbourhoodId"))
             .withPrice(rs.getFloat("price"))
             .withPrivacyLevel(rs.getBoolean("privacyLevel"))
             .withPropertyType(PropertyType.valueOf(rs.getInt("propertyType")))
